# docs/swagger.yaml
definitions:
...
  models.CustomerBase:
    properties:
      additional_info:
        additionalProperties:
          type: string
        type: object
      address:
        description: Email   string `json:"email" bson:"email"`
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  models.Error:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.Output:
    properties:
      data: {}
      error:
        items:
          $ref: "#/definitions/models.Error"
        type: array
    type: object
...
info:
  contact:
    email: hamidovaslon13@gmail.com
    name: API Support
    url: https://github.com/aslon1213/g4h_pos_erp
  description: This is a ERP/POS API for Magazin.
  title: Magazin ERP/POS API
  version: "1.0"
paths:
...
  /api/customers:
    get:
      consumes:
        - application/json
      description: Get all customers from the database
      parameters:
        - description: Customer name
          in: query
          name: name
          type: string
        - description: Customer phone
          in: query
          name: phone
          type: string
        - description: Customer address
          in: query
          name: address
          type: string
        - description: Page number
          in: query
          name: page
          type: integer
        - description: Number of customers per page
          in: query
          name: count
          type: integer
        - description: Sort by BNPL total (max, min, none)
          in: query
          name: sort_by_bnpl_total
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.Output"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.Output"
      security:
        - BearerAuth: []
      summary: Get all customers
      tags:
        - customers
    post:
      consumes:
        - application/json
      description: Create a new customer in the database
      parameters:
        - description: Customer data
          in: body
          name: customer
          required: true
          schema:
            $ref: "#/definitions/models.CustomerBase"
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/models.Output"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/models.Output"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.Output"
      security:
        - BearerAuth: []
      summary: Create a new customer
      tags:
        - customers
...
  /api/customers/{id}:
    delete:
      consumes:
        - application/json
      description: Delete a customer from the database
      parameters:
        - description: Customer ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.Output"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/models.Output"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.Output"
      security:
        - BearerAuth: []
      summary: Delete a customer
      tags:
        - customers
    get:
      consumes:
        - application/json
      description: Get a customer by its ID
      parameters:
        - description: Customer ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.Output"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/models.Output"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.Output"
      security:
        - BearerAuth: []
      summary: Get a customer by ID
      tags:
        - customers
    put:
      consumes:
        - application/json
      description: Update an existing customer in the database
      parameters:
        - description: Customer ID
          in: path
          name: id
          required: true
          type: string
        - description: Customer data
          in: body
          name: customer
          required: true
          schema:
            $ref: "#/definitions/models.CustomerBase"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.Output"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/models.Output"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/models.Output"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.Output"
      security:
        - BearerAuth: []
      summary: Update a customer
      tags:
        - customers
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"